Generalize scripts and make them configurable for reuse. (#1)
Making the scripts as generic as possible is essential to emulate incorrectly configured clusters and simplify the fixing process. An operator can reuse them in production environments to speed up the upgrading and configuration process. Besides generalization, readability is necessary, including commenting to ensure an operator understands what the scripts do.